% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_assignment_status}
\alias{plot_assignment_status}
\title{Plot assignment status counts}
\usage{
plot_assignment_status(res)
}
\arguments{
\item{res}{Result of build_pedigree()}
}
\description{
Plot assignment status counts
}
