% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_kinship}
\alias{plot_kinship}
\title{Plot candidate-parent similarity distribution}
\usage{
plot_kinship(res)
}
\arguments{
\item{res}{Result of build_pedigree()}
}
\description{
Plot candidate-parent similarity distribution
}
