% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_me}
\alias{plot_me}
\title{Plot Mendelian error rate histogram for tested trios}
\usage{
plot_me(res)
}
\arguments{
\item{res}{Result of build_pedigree()}
}
\description{
Plot Mendelian error rate histogram for tested trios
}
