% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_outputs.R
\name{write_outputs}
\alias{write_outputs}
\title{Write geno2ped outputs to disk}
\usage{
write_outputs(res, outdir = "geno2ped_results", prefix = NULL)
}
\arguments{
\item{res}{Result from \code{build_pedigree()}}

\item{outdir}{Output directory}

\item{prefix}{Optional file prefix}
}
\value{
Invisibly returns a character vector of written file paths
}
\description{
Write geno2ped outputs to disk
}
